requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-webpack-loader==0.7.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
